+++ src/main/java/nextstep/ladder/LadderMain.java
@@ -1,5 +1,12 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Ladder;
+import nextstep.ladder.domain.Person;
+import nextstep.ladder.factory.LadderFactory;
+import nextstep.ladder.factory.RandomLadderFactory;
+import nextstep.ladder.view.InputView;
+import nextstep.ladder.view.OutputView;
+
 public class LadderMain {
     public static void main(String[] args) {
         Person person = new Person(InputView.inputPerson());
