+++ src/main/java/nextstep/ladder/factory/LadderFactory.java
@@ -1,4 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.factory;
+
+import nextstep.ladder.domain.Ladder;
 
 public interface LadderFactory {
     Ladder creatLadder(int height, int countPerson);
