+++ src/test/java/nextstep/ladder/LadderFactoryTest.java
@@ -1,5 +1,8 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Ladder;
+import nextstep.ladder.factory.LadderFactory;
+import nextstep.ladder.factory.RandomLadderFactory;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
