+++ src/test/java/nextstep/ladder/PersonTest.java
@@ -1,5 +1,6 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Person;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
